split up main plugin file & fix + add tests
