+++ tests/Functional/RestApiCest.php
@@ -215,14 +215,14 @@ final class RestApiCest {
     /**
      * Test that showing a private post in the rest api as an editor user with the context "edit" does not expose the key.
      */
-    public function try_to_not_see_key_for_private_post_in_authenticated_request_with_edit_context( FunctionalTester $I ): void {
+    public function try_to_see_key_for_private_post_in_authenticated_request_with_edit_context( FunctionalTester $I ): void {
         $post_id = $I->havePrivatePostWithSharableLinkInDatabase(
             [
                 'post_title' => 'Private Post',
                 'post_content' => 'private content',
             ]
         );
-        $key = $I->grabPostMetaFromDatabase( $post_id, '_sppp_key', true );
+        $key = $I->grabSecretKeyFromDatabase( $post_id );
 
         $this->loginAsEditor( $I );
 
@@ -230,23 +230,23 @@ final class RestApiCest {
         $I->seeResponseCodeIs( 200 );
         $response = $I->grabResponse();
         $I->assertArrayHasKey( 'meta', $response );
-        $I->assertArrayHasKey( '_sppp_enabled', $response['meta'] );
-        $I->assertTrue( $response['meta']['_sppp_enabled'] );
+        $I->assertArrayHasKey( '_sppp_enabeled', $response['meta'] );
         $I->assertArrayHasKey( '_sppp_key', $response['meta'] );
+        $I->assertTrue( $response['meta']['_sppp_enabeled'] );
         $I->assertEquals( $response['meta']['_sppp_key'], $key );
     }
 
     /**
      * Test that showing a private post in the rest api as an editor user with the context "edit" does not expose the key.
      */
-    public function try_to_not_see_key_for_password_protected_post_in_authenticated_request_with_edit_context( FunctionalTester $I ): void {
+    public function try_to_see_key_for_password_protected_post_in_authenticated_request_with_edit_context( FunctionalTester $I ): void {
         $post_id = $I->havePasswordProtectedPostWithSharableLinkInDatabase(
             [
                 'post_title' => 'Password protected Post',
                 'post_content' => 'private content',
             ]
         );
-        $key = $I->grabPostMetaFromDatabase( $post_id, '_sppp_key', true );
+        $key = $I->grabSecretKeyFromDatabase( $post_id );
 
         $this->loginAsEditor( $I );
 
@@ -255,8 +255,9 @@ final class RestApiCest {
         $response = $I->grabResponse();
         $I->assertArrayHasKey( 'meta', $response );
         $I->assertArrayHasKey( '_sppp_enabled', $response['meta'] );
-        $I->assertTrue( $response['meta']['_sppp_enabled'] );
         $I->assertArrayHasKey( '_sppp_key', $response['meta'] );
+
+        $I->assertTrue( $response['meta']['_sppp_enabled'] );
         $I->assertEquals( $response['meta']['_sppp_key'], $key );
     }
 
@@ -285,7 +286,7 @@ final class RestApiCest {
         $I->seeResponseCodeIs( 200 );
         $response = $I->grabResponse();
         $I->assertArrayHasKey( 'meta', $response );
-        $I->assertArrayHasKey( '_sppp_enabled', $response['meta'] );
+        $I->assertArrayNotHasKey( '_sppp_enabled', $response['meta'] );
         $I->assertTrue( $response['meta']['_sppp_enabled'] );
         $I->assertArrayHasKey( '_sppp_key', $response['meta'] );
         $I->assertEquals( $response['meta']['_sppp_key'], $key );
