+++ tests/Unit/_bootstrap.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * Unit suite bootstrap file.
+ *
+ * Here you can initialize variables that will be available to your tests
+ */
+
+// require_once dirname(__DIR__, 2) . '/vendor/php-stubs/wordpress-stubs/wordpress-stubs.php';
+
+require __DIR__ . '/../../vendor/vendor-prefixed/autoload.php';
